+++ avantpy/converter.py
@@ -37,15 +37,11 @@ which might be converted into::
 
     variable =function(argument)
 
-One possibility that AvantPy offers is to run program with the
---diff option to show the difference
-between the code written and standard Python. The difflib
-module, used to do this, shows all differences including
-differences in spaces between tokens.  Since the primary
-goal of the --diff option is to allow users to learn the
-differences between their dialect and standard Python, it is
-important to restrict differences shown to only those that
-are meaningful.  As a result, we do not use Python's
+This could make it more difficult to compare the original
+code with the converted one, as it is possible to do
+using one of the utilises provided with AvantPy,
+or any "diff" program.
+As a result, we do not use Python's
 untokenize function, and explicitly keep track of spacing
 between tokens.
 
@@ -222,15 +218,11 @@ class Converter:
 
         variable =function(argument)
 
-    One possibility that AvantPy offers is to run program with the
-    --diff option to show the difference
-    between the code written and standard Python. The difflib
-    module, used to do this, shows all differences including
-    differences in spaces between tokens.  Since the primary
-    goal of the --diff option is to allow users to learn the
-    differences between their dialect and standard Python, it is
-    important to restrict differences shown to only those that
-    are meaningful.  As a result, we do not use Python's
+    This could make it more difficult to compare the original
+    code with the converted one, as it is possible to do
+    using one of the utilises provided with AvantPy,
+    or any "diff" program.
+    As a result, we do not use Python's
     untokenize function, and explicitly keep track of spacing
     between tokens."""