+++ avantpy/exceptions.py
@@ -16,38 +16,6 @@ class IfNobreakError(AvantPyException):
     pass
 
 
-class TryNobreakError(AvantPyException):
-    """Raised if ``nobreak`` is used instead of ``else`` in ``try/except`` statement"""
-
-    pass
-
-
-class NobreakFirstError(AvantPyException):
-    """Raised if ``nobreak`` is used somewhere other than at the beginning of a line."""
-
-    pass
-
-
-class NobreakSyntaxError(AvantPyException):
-    """Raised if ``nobreak`` is without a matching block."""
-
-    pass
-
-
-class RepeatFirstError(AvantPyException):
-    """Raised if ``repeat`` is used somewhere other than at the beginning of a line."""
-
-    pass
-
-
-class MissingRepeatError(AvantPyException):
-    """Raised if ``until`` or ``forever`` is used without being preceeded by
-       ``repeat``.
-    """
-
-    pass
-
-
 class MismatchedBracketsError(AvantPyException):
     """Raised if a left parenthesis '(', or square bracket '['
        or curly bracket '{', is closed by a right 'bracket' of a different type.
@@ -73,6 +41,21 @@ class MissingLeftBracketError(AvantPyException):
     pass
 
 
+class MissingRepeatError(AvantPyException):
+    """Raised if ``until`` or ``forever`` is used without being preceeded by
+       ``repeat``.
+    """
+
+    pass
+
+
+class MissingRepeatColonError(AvantPyException):
+    """Raised if a line beginning with ``repeat`` does not end with a colon.
+    """
+
+    pass
+
+
 class MissingRightBracketError(AvantPyException):
     """Raised if a left parenthesis '(', or square bracket '['
        or curly bracket '{', is never closed by a corresponding bracket.
@@ -85,6 +68,30 @@ class MissingRightBracketError(AvantPyException):
     pass
 
 
+class NobreakFirstError(AvantPyException):
+    """Raised if ``nobreak`` is used somewhere other than at the beginning of a line."""
+
+    pass
+
+
+class NobreakSyntaxError(AvantPyException):
+    """Raised if ``nobreak`` is without a matching block."""
+
+    pass
+
+
+class RepeatFirstError(AvantPyException):
+    """Raised if ``repeat`` is used somewhere other than at the beginning of a line."""
+
+    pass
+
+
+class TryNobreakError(AvantPyException):
+    """Raised if ``nobreak`` is used instead of ``else`` in ``try/except`` statement"""
+
+    pass
+
+
 class UnknownDialectError(AvantPyException):
     """Raised when attempting to set ``dialect`` to unsupported value."""
 
@@ -93,12 +100,5 @@ class UnknownDialectError(AvantPyException):
 
 class UnknownLanguageError(AvantPyException):
     """Raised when attempting to set ``lang`` to unsupported value."""
-
-    pass
-
-
-class UnexpectedError(AvantPyException):
-    """Raised when some unexpected condition occurs which should
-       be cavered by a specific excption."""
 
     pass