+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- Simplified the work required to create documentation for simplified tracebacks in a given language
+- Wrote a script to duplicate all tests for pyupper in the other dialects.
+- Added full test for pyupper (and other dialects)
 - Changed translations so that they are done with gettext
 - Removed old way of supporting translations of human languages
 - Printing information about language and dialect used in console